Collapse integration tests into single execution
