+++ .github/workflows/integration-test.yml
@@ -47,14 +47,6 @@ jobs:
       - name: Setup Cloud SDK
         uses: google-github-actions/setup-gcloud@v1
 
-      - name: Generate Encryption Passphrase
-        run: openssl rand -base64 32
-        if: ${{ false }}
-
-      - name: Encrypt Google Services config file
-        if: ${{ false }}
-        run: find . -type f -name "google-services.json" -exec sh -c 'openssl aes-256-cbc -md sha256 -salt -in "$0" -out "${0%.json*}.aes" -k "${GOOGLE_SERVICES_PASSPHRASE}"' {} \;
-
       - name: Decrypt Google Services
         if: ${{ !contains(github.event.pull_request.labels.*.name, 'Dry Run') }}
         env:
@@ -75,8 +67,8 @@ jobs:
           arguments: |
             cloud-run:integrationTest
               ${{ (contains(github.event.pull_request.labels.*.name, 'Dry Run') && '--dry-run' || '') }}
+              ${{ inputs.debug && '--debug' || '' }}
               --console=plain
-              --info
 
       - name: Gradle Scan Link
         uses: actions/github-script@v6
