--- cloud-run/src/jvmIntegrationTest/kotlin/io/ashdavies/cloud/AuthTest.kt
@@ -1,28 +1,0 @@
-package io.ashdavies.cloud
-
-import io.ktor.client.call.body
-import io.ktor.client.request.header
-import io.ktor.client.request.post
-import io.ktor.client.request.setBody
-import io.ktor.http.ContentType
-import io.ktor.http.contentType
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlin.test.Test
-import kotlin.test.assertNotNull
-
-@ExperimentalCoroutinesApi
-internal class AuthTest {
-
-    @Test
-    fun `should sign in with custom token`() = testMainApplication { client ->
-        val apiKey = requireNotNull(System.getenv("GOOGLE_PROJECT_API_KEY"))
-
-        val authResult = client.post("/firebase/auth") {
-            setBody(mapOf("uid" to "jane.smith@example.com"))
-            contentType(ContentType.Application.Json)
-            header("X-API-Key", apiKey)
-        }.body<Map<String, String>>()
-
-        assertNotNull(authResult["idToken"])
-    }
-}
